WIP migration to shared IrcClientCore for connections
The shared IrcClientCore is something I've been working on for a couple years, and has been the base of an ASP.Net based Web IRC client I've been working on but have yet to open source as it's still in early development. The main changes compared to the winirc code it was based off is a much better seperation of the UI from the logic, and a much more modular implementation of the irc line handler and the commands used by the user.

Right now there's still things that are broken, including logging to files, however getting WinIRC back to feature parity is my next priority. I'm also gonna get a mirror of the IrcClientCore on GitHub ASAP, with an intent to move development fully to GitHub once I figure out a replacement to the GitLab CI currently building a new NuGet package on each release